LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-back-after-pushState-expected.txt


FAIL history.back() fires the navigate event when reversing a pushState() assert_equals: expected "traverse" but got "push"
